게임 클리어 UI  추가 ,임시 한글 폰트 추가, 보스 골드 추가,보스 Dead상태 추가, 보스 처치 시 게임 클리어 UI 표시

+++ Assets/01.Scripts/Monsters/Monster.cs
@@ -14,7 +14,7 @@ public class Monster : LivingEntity, IPlayerObserver
 
     public float Damage { get; private set; }
     public float Exp { get; private set; }
-    public float Gold { get; private set; }
+    public int Gold { get; private set; }
     public float TotalCooldown { get; private set; }
     public float Range { get; private set; }
